+++ drivers/gpu/nvgpu/common/vgpu/debugger_vgpu.h
@@ -32,7 +32,7 @@ struct channel_gk20a;
 int vgpu_exec_regops(struct gk20a *g,
 		     struct channel_gk20a *ch,
 		     struct nvgpu_dbg_reg_op *ops,
-		     u64 num_ops,
+		     u32 num_ops,
 		     bool is_profiler,
 		     bool *is_current_ctx);
 int vgpu_dbg_set_powergate(struct dbg_session_gk20a *dbg_s,
